// src/app/shared/models/buildingType.ts
export enum BuildingType {
  CRYSTAL_MINE = 'CRYSTAL_MINE',
  DEUTERIUM_SYNTHESIZER = 'DEUTERIUM_SYNTHESIZER',
  METAL_MINE = 'METAL_MINE',
  FUSION_REACTOR = 'FUSION_REACTOR',
  SOLAR_PLANT = 'SOLAR_PLANT',
  SOLAR_SATELLITE = 'SOLAR_SATELLITE',
  CRYSTAL_STORAGE = 'CRYSTAL_STORAGE',
  METAL_STORAGE = 'METAL_STORAGE',
  DEUTERIUM_TANK = 'DEUTERIUM_TANK',
  ROBOTICS_FACTORY = 'ROBOTICS_FACTORY',
  SHIPYARD = 'SHIPYARD',
  RESEARCH_LAB = 'RESEARCH_LAB',
  ALLIANCE_DEPOT = 'ALLIANCE_DEPOT',
  MISSILE_SILO = 'MISSILE_SILO',
  NANITE_FACTORY = 'NANITE_FACTORY',
  TERRAFORMER = 'TERRAFORMER',
  // technologies
  COMBUSTION_DRIVE = 'COMBUSTION_DRIVE',
  IMPULSE_DRIVE = 'IMPULSE_DRIVE',
  ARMOUR_TECHNOLOGY = 'ARMOUR_TECHNOLOGY',
  ION_TECHNOLOGY = 'ION_TECHNOLOGY',
  HYPERSPACE_TECHNOLOGY = 'HYPERSPACE_TECHNOLOGY',
  HYPERSPACE_DRIVE = 'HYPERSPACE_DRIVE',
  LASER_TECHNOLOGY = 'LASER_TECHNOLOGY',
  PLASMA_TECHNOLOGY = 'PLASMA_TECHNOLOGY',
  GRAVITON_TECHNOLOGY = 'GRAVITON_TECHNOLOGY',
  SHIELDING_TECHNOLOGY = 'SHIELDING_TECHNOLOGY',
  ESPIONAGE_TECHNOLOGY = 'ESPIONAGE_TECHNOLOGY',
  COMPUTER_TECHNOLOGY = 'COMPUTER_TECHNOLOGY',
  WEAPON_TECHNOLOGY = 'WEAPON_TECHNOLOGY',
  ENERGY_TECHNOLOGY = 'ENERGY_TECHNOLOGY',
  INTERGALACTIC_RESEARCH_NETWORK = 'INTERGALACTIC_RESEARCH_NETWORK',
  ASTROPHYSICS = 'ASTROPHYSICS',
}